Remove duplicate PhotonClientStore from elements
