+++ packages/elements/src/photon-login/photon-login-component.tsx
@@ -1,7 +1,7 @@
 import '../photon-auth-button';
 import { customElement } from 'solid-element';
 import { createEffect } from 'solid-js';
-import { usePhoton } from '../context';
+import { usePhoton } from '@photonhealth/components';
 import tailwind from '../tailwind.css?inline';
 
 const Component = (props: { redirectPath?: string }) => {
