+++ packages/elements/src/photon-patient-select/photon-patient-select-component.tsx
@@ -2,7 +2,7 @@
 import { customElement } from 'solid-element';
 
 //Photon
-import { usePhoton } from '../context';
+import { usePhoton } from '@photonhealth/components';
 import { PhotonDropdown } from '../photon-dropdown';
 
 //Types
